pyproject.toml
[project]
name = "topic-analysis-experiments"
version = "0.1.0"
description = "Statistical Topic Analysis for Medical Education Corpus"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.0.0",
    "pandas>=2.2.0",
    "gensim>=4.3.0",
    "bertopic>=0.16.0",
    "sentence-transformers>=2.6.0",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.6",
    "scikit-learn>=1.4.0",
    "scipy>=1.11.0",
    "nltk>=3.8.0",
    "spacy>=3.7.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "typer>=0.12.0",
]

# Note: MANTA (manta-topic-modelling) requires older scipy/scikit-learn
# Install separately if needed: pip install manta-topic-modelling

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/topic_analysis"]
